updated folders

+++ utils/prediction_handeling.py
@@ -2,6 +2,6 @@ import numpy as np
 
 def handle_negative_predictions(y):
 
-    y = np.array(y, dtype=float)
-    y[y < 0] = 0.0
+    y = np.array(y, dtype=int)
+    y[y < 0] = 0
     return y
